refactor: remove redundant service implementations
